Refactoring Queue Handler
