--- queue-manager/src/main/scala/it/pagopa/interop/commons/queue/impl/SQSDequeuedMessage.scala
@@ -1,3 +1,0 @@
-package it.pagopa.interop.commons.queue.impl
-
-final case class SQSDequeuedMessage[V](value: V, receiptHandle: String)
